+++ InventoryManagementSystem/src/main/java/com/example/inventorymanagementsystem/repositories/ItemRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ItemRepository extends MongoRepository<Item,Integer> {
+public interface ItemRepository extends MongoRepository<Item,Long> {
 
-    Optional<Item> findByItemIdAndProductId(Integer itemId, Integer productId);
+    Optional<Item> findByItemIdAndProductId(Long itemId, Integer productId);
 }
